Update partitioning scheme to put swap first
The default layout of the atomic scheme is that swap comes second.
This becomes a problem is when provisioning on large VM disks as the
swap partition sits between the root partition and the free space.

Putting the root after the swap, it is now easy to grow the root to the
available space.

Also updated is the disk image size.  This could be shrunk